@@ -28,7 +28,7 @@
   "devDependencies": {
     "babel-eslint": "^10.0.1",
     "enzyme": "^3.7.0",
-    "eslint": "^5.6.1",
+    "eslint": "^5.7.0",
     "eslint-plugin-react": "^7.11.1",
     "jest": "^23.6.0",
     "react-test-renderer": "^16.5.2",
