@@ -166,7 +166,7 @@ eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nvar StateApi = function () {\n  function StateApi(rawData) {\n    var _this = this;\n\n    _classCallCheck(this, StateApi);\n\n    this.lookupAuthor = function (authorId) {\n      return _this.data.authors[authorId];\n    };\n\n    this.getState = function () {\n      return _this.data;\n    };\n\n    this.data = {\n      articles: this.mapIntoObject(rawData.articles),\n      authors: this.mapIntoObject(rawData.authors)\n    };\n  }\n\n  _createClass(StateApi, [{\n    key: \"mapIntoObject\",\n    value: function mapIntoObject(arr) {\n      return arr.reduce(function (acc, curr) {\n        acc[curr.id] = curr;\n        return acc;\n      }, {});\n    }\n  }]);\n\n  return StateApi;\n}();\n\nexports.default = StateApi;//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9saWIvc3RhdGUtYXBpL2xpYi9pbmRleC5qcy5qcyIsInNvdXJjZXMiOlsid2VicGFjazovLy9saWIvc3RhdGUtYXBpL2xpYi9pbmRleC5qcz82YWM1Il0sInNvdXJjZXNDb250ZW50IjpbImNsYXNzIFN0YXRlQXBpIHtcclxuICBjb25zdHJ1Y3RvcihyYXdEYXRhKSB7XHJcbiAgICB0aGlzLmRhdGEgPSB7XHJcbiAgICAgIGFydGljbGVzOiB0aGlzLm1hcEludG9PYmplY3QocmF3RGF0YS5hcnRpY2xlcyksXHJcbiAgICAgIGF1dGhvcnM6IHRoaXMubWFwSW50b09iamVjdChyYXdEYXRhLmF1dGhvcnMpLFxyXG4gICAgfTtcclxuICB9XHJcbiAgbWFwSW50b09iamVjdChhcnIpIHtcclxuICAgIHJldHVybiBhcnIucmVkdWNlKChhY2MsIGN1cnIpID0+IHtcclxuICAgICAgYWNjW2N1cnIuaWRdID0gY3VycjtcclxuICAgICAgcmV0dXJuIGFjYztcclxuICAgIH0sIHt9KTtcclxuICB9XHJcbiAgbG9va3VwQXV0aG9yID0gKGF1dGhvcklkKSA9PiB7XHJcbiAgICByZXR1cm4gdGhpcy5kYXRhLmF1dGhvcnNbYXV0aG9ySWRdO1xyXG4gIH1cclxuICBnZXRTdGF0ZSA9ICgpID0+IHtcclxuICAgIHJldHVybiB0aGlzLmRhdGE7XHJcbiAgfVxyXG59XHJcblxyXG5leHBvcnQgZGVmYXVsdCBTdGF0ZUFwaTtcclxuIl0sIm1hcHBpbmdzIjoiOzs7Ozs7Ozs7O0FBQUE7QUFDQTtBQUFBO0FBQ0E7QUFEQTtBQUNBO0FBREE7QUFhQTtBQUNBO0FBQ0E7QUFmQTtBQWdCQTtBQUNBO0FBQ0E7QUFqQkE7QUFDQTtBQUNBO0FBRkE7QUFJQTtBQUNBOzs7QUFBQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7Ozs7OztBQVNBIiwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./lib/state-api/lib/index.js\n");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nvar StateApi = function () {\n  function StateApi(rawData) {\n    var _this = this;\n\n    _classCallCheck(this, StateApi);\n\n    this.lookupAuthor = function (authorId) {\n      return _this.data.authors[authorId];\n    };\n\n    this.getState = function () {\n      return _this.data;\n    };\n\n    this.data = {\n      articles: this.mapIntoObject(rawData.articles),\n      authors: this.mapIntoObject(rawData.authors),\n      searchTerm: ''\n    };\n  }\n\n  _createClass(StateApi, [{\n    key: 'mapIntoObject',\n    value: function mapIntoObject(arr) {\n      return arr.reduce(function (acc, curr) {\n        acc[curr.id] = curr;\n        return acc;\n      }, {});\n    }\n  }]);\n\n  return StateApi;\n}();\n\nexports.default = StateApi;//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiLi9saWIvc3RhdGUtYXBpL2xpYi9pbmRleC5qcy5qcyIsInNvdXJjZXMiOlsid2VicGFjazovLy9saWIvc3RhdGUtYXBpL2xpYi9pbmRleC5qcz82YWM1Il0sInNvdXJjZXNDb250ZW50IjpbImNsYXNzIFN0YXRlQXBpIHtcclxuICBjb25zdHJ1Y3RvcihyYXdEYXRhKSB7XHJcbiAgICB0aGlzLmRhdGEgPSB7XHJcbiAgICAgIGFydGljbGVzOiB0aGlzLm1hcEludG9PYmplY3QocmF3RGF0YS5hcnRpY2xlcyksXHJcbiAgICAgIGF1dGhvcnM6IHRoaXMubWFwSW50b09iamVjdChyYXdEYXRhLmF1dGhvcnMpLFxyXG4gICAgICBzZWFyY2hUZXJtOiAnJyxcclxuICAgIH07XHJcbiAgfVxyXG4gIG1hcEludG9PYmplY3QoYXJyKSB7XHJcbiAgICByZXR1cm4gYXJyLnJlZHVjZSgoYWNjLCBjdXJyKSA9PiB7XHJcbiAgICAgIGFjY1tjdXJyLmlkXSA9IGN1cnI7XHJcbiAgICAgIHJldHVybiBhY2M7XHJcbiAgICB9LCB7fSk7XHJcbiAgfVxyXG4gIGxvb2t1cEF1dGhvciA9IChhdXRob3JJZCkgPT4ge1xyXG4gICAgcmV0dXJuIHRoaXMuZGF0YS5hdXRob3JzW2F1dGhvcklkXTtcclxuICB9XHJcbiAgZ2V0U3RhdGUgPSAoKSA9PiB7XHJcbiAgICByZXR1cm4gdGhpcy5kYXRhO1xyXG4gIH1cclxufVxyXG5cclxuZXhwb3J0IGRlZmF1bHQgU3RhdGVBcGk7XHJcbiJdLCJtYXBwaW5ncyI6Ijs7Ozs7Ozs7OztBQUFBO0FBQ0E7QUFBQTtBQUNBO0FBREE7QUFDQTtBQURBO0FBY0E7QUFDQTtBQUNBO0FBaEJBO0FBaUJBO0FBQ0E7QUFDQTtBQWxCQTtBQUNBO0FBQ0E7QUFDQTtBQUhBO0FBS0E7QUFDQTs7O0FBQUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBOzs7Ozs7QUFTQSIsInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./lib/state-api/lib/index.js\n");
 
 /***/ }),
 
