@@ -1,6 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import pickBy from 'lodash.pickby';
 
 import ArticleList from './ArticleList';
 import SearchBar from './SearchBar';
@@ -22,18 +21,11 @@ setSearchTerm = (searchTerm) => {
   this.setState({ searchTerm});
 }
 render() {
-  let { articles, searchTerm} = this.state;
-  if (searchTerm) {
-    articles = pickBy(articles, (key, value) => {
-      return value.title.match(searchTerm)
-        || value.body.match(searchTerm);
-    });
-  }
   return (
     <div>
       <SearchBar doSearch={this.setSearchTerm} />,
       <ArticleList
-        articles={articles}
+        articles={this.state.articles}
         store={this.props.store}
       />
     </div>
